requirements.txt
python-telegram-bot==21.*
elevenlabs==1.*
python-dotenv==1.*
pydantic==2.*
pydantic-settings==2.*
pytest==8.*
httpx==0.27.*
